+++ Selu383.SP24.Web/src/Components/HotelDto.tsx
@@ -4,7 +4,7 @@ export interface HotelDto {
 	address: string;
 	city: string;
 	state: string;
-	postalcode: string;
+	postalCode: string;
 	managerId: number;
 }
