+++ Selu383.SP24.Web/tsconfig.json
@@ -5,6 +5,9 @@
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "module": "ESNext",
     "skipLibCheck": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "types": ["node"],
 
     /* Bundler mode */
     "moduleResolution": "bundler",
